* commit '3b53d26cf793ac89d6420fe36b1c06bb1596b64d': (129 commits)
  change version
  加油正式环境1分钱
  change version
  temp ci
  temp ci
  change version
  change version
  保险支付返回订单详情
  ci
  保险核保结果顶部的tips
  保险一个cell重用
  支付
  change version
  ci
  change version
  ci
  ci
  ci
  ci
  优惠券去使用优化和bug修改。保险支付协议缺失bug修改。省钱攻略替换为detailwebvc。
  ...

+++ XiaoMa/Codemap/netop/GascardChargeOp.h
@@ -7,7 +7,7 @@
 ///优惠劵
 @property (nonatomic,strong) NSNumber* req_cid;
 ///充值金额
-@property (nonatomic,assign) int req_amount;
+@property (nonatomic,assign) NSInteger req_amount;
 ///支付方式
 @property (nonatomic,assign) int req_paychannel;
 ///支付验证码
@@ -22,9 +22,9 @@
 ///记录ID
 @property (nonatomic,strong) NSNumber* rsp_orderid;
 ///支付金额
-@property (nonatomic,assign) int rsp_total;
+@property (nonatomic,assign) CGFloat rsp_total;
 ///优惠金额
-@property (nonatomic,assign) int rsp_couponmoney;
+@property (nonatomic,assign) CGFloat rsp_couponmoney;
 
 
 @end
